modified table formatting
