+++ client/public/index.html
@@ -25,10 +25,11 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <title>SimplExpend</title>
+    <script src="https://kit.fontawesome.com/97e4b01de0.js" crossorigin="anonymous"></script>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-
+    
   </body>
 </html>
